Change entity class structure
