+++ entity/model/base.py
